Revert "verifier: Refactor type verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1a75347-c078-4c1f-954e-a9b27f4e19e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15718cf6d5ef6c9e80d6faa85b34f73007739994
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com