Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a2bc0fe-5b60-4714-b8a0-952b67e91d90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5d8b2534ad4cbe4a166aa23252d12e891ac64b03
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com