Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef5c57b3-75cf-47e5-99a2-b80857aa7cda). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c93e6a6731384e187aea8074c1867f9e0356fe5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com