Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d7a6996-0088-4fb8-b8c2-d6fc1fe7cba3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b53406386cac9c262b8d093df1f071103332d36
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com