Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/064f055c-ff6f-4f03-acb8-b67b87939c16). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id8156147e3d110df3855ed0c51a3e07fba2b94fc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com