Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3504a06b-4eae-4455-8e47-24bd1137e1f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4405e6b9467c7881910a82fa34ed6913c5a94347
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com