Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ca76725-d344-418c-9ba5-83ab6420d4d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07acac727515817d913297ecd73c089d6a17d5c5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com