verifier: Fill merge table for `Constant{Lo,Hi}`.
The code for these cases was removed in the
    https://android-review.googlesource.com/3342468
rewrite because it became an unnecessaary run-time check
since `RegType::Merge()` is never called with such kinds,
see the `DCHECK()` at the start. Now that this code is
used at compile time for generating a pre-compiled table,
we can fill these unused entries with no run-time cost.
The `RegisterLine::MergeRegister()` could be reimplemented
in a way that would actually use these entries.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: I482a3ae9756f576561a86f256d63f84537ea4fc5