Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b3f31bc-5e62-483d-956d-fe0b81669ab7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I19980c28ec0d20df34b2c57001fcdd3f8c4fcef6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com