Revert "Adjust art tests to java.lang changes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/560f6367-4ab9-4832-a318-9aa6e6bbfa5e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3bdd458bc50c98626e9207a6d9eba59e6be1f0a4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com