Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/032588e9-1022-42c7-a6bc-642723b5d963). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5e8bda7f1f478a9f5825cada3c273001776a176
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com