Revert "Add poisoning/unpoisoning to invokeExact intrinsic."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2639f551-7598-4dab-895f-82e34ce5b66c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b22a905621a793a06d55dde39d62316944a29ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com