Revert "Refactor ResolveMethod."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/084376c5-4709-4d0e-989c-d940e161ac80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I52e8a71d44e4678ab701983767ff78be8ff2fbb8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com