Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a71ed345-042b-49ac-9046-3fefef7761c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8656da2f04ddb6f5a8c3fef8dece45815cf7ea8e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com