Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e273fdc4-7504-4857-a2f3-158ffa0ab640). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id39d39a064186e8acf7e6ce11687266cef0e7ba2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com