Revert "Add a dump option to dump SDM status."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d83d69f1-1a30-4e78-b5ea-392e1e91c17c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f18582fc1888c3f89c70b155d7e9ddb14a47ce4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com