Revert "verifier: Faster primitive assignability checks."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ee60e65-e446-4d49-84b9-fd4316059d50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie10c4de8d075a045ba91a1fa5b97b3835fc6fc6b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com