Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb82f764-53ab-4df8-9b94-9096bd6b0c02). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88d3d78a77939afafbcc3f85a64534b850bae4b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com