Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eeb8a8e6-4f8f-4086-ab46-7ac37765a2a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie39c1b4fc9f98b43abd0bbcaa9518a1e6670ed72
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com